+++ pkg/meta/BUILD.bazel
@@ -17,8 +17,8 @@ go_library(
         "//pkg/parser/mysql",
         "//pkg/structure",
         "//pkg/util/dbterror",
-        "//pkg/util/fastjson",
         "//pkg/util/hack",
+        "//pkg/util/partialjson",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_kvproto//pkg/kvrpcpb",
         "@com_github_pingcap_kvproto//pkg/resource_manager",
